requirements.txt
Django>=5.0,<6.0
django-environ>=0.11
python-dateutil>=2.9
pytz>=2024.1
whitenoise>=6.7
gunicorn>=23.0; platform_system != "Windows"
# Dev
pytest>=8.2
pytest-django>=4.9
ruff>=0.6
black>=24.8